dnetview/ parse: make NodeId unique to each node
solve problem of endless creation of new node_ids. we derive the id from
the node_name, insuring there are no duplicates.

our method for doing this right now is very weird- we convert the String
node_name into a u64. going forward we will likely replace the node_id
with a hash to be more straightforward.