runtime/merkle: Introduce pointers to latest merkle roots in merkle_add()
Also migrate all the contracts that use merkle_add to this API.

This is the better approach, since it avoids having to run redundant
operations like lots of hashing when we want to retrieve the latest
Merkle root in some tree. It is too expensive to do it directly in WASM,
so doing this through the injected host function allows us to price it
more accordingly.