explorer: add no-sync mode to run explorerd without connecting or syncing with darkfid
Introduced a `no-sync` mode to the `explorerd` daemon, allowing the application to operate solely on a local explorer database without connecting to a `darkfid` blockchain network or syncing new blocks. This feature is particularly
useful for UI development, testing, development tasks that don't require syncing, or exploring functionality without the overhead of running a blockchain node or miner.

This setup reduces resource consumption and accelerates the development workflow by avoiding the need to sync or connect to a blockchain network when it is not needed. It also simplifies the build setup, as the explorer can run without building `darkfid` or `minerd`, reducing dependencies.

Updates include:
- Added a `--no-sync` CLI argument to toggle this mode
- Updated the sync logic to skip tasks like `subscriber_task` and `listener_task` when `no-sync` mode is enabled
- Enhanced the `Explorerd` struct to support conditional connection handling
- Updated the startup banner to display `(No-Sync Mode)` and to show "Not connected" for the connected node in `no-sync` mode